MOD: Renaming and moving of LookNFeelPropertyEditorDockWidget

+++ ceed/editors/looknfeel/visual.py
@@ -22,12 +22,10 @@
 #   along with this program.  If not, see <http://www.gnu.org/licenses/>.
 ##############################################################################
 
-from collections import OrderedDict
 
 from PySide import QtCore
 from PySide import QtGui
 import cPickle
-import os
 
 import PyCEGUI
 
@@ -37,17 +35,12 @@ from ceed.editors import multi
 
 from ceed.cegui import widgethelpers as cegui_widgethelpers
 
-from ceed.editors.looknfeel import undo
+from ceed.editors.looknfeel import undoable_commands
 from ceed.editors.looknfeel import widgethelpers
 
-from ceed.propertysetinspector import PropertyInspectorWidget
-from ceed.propertysetinspector import CEGUIWidgetLookPropertiesManager
-
-import ceed.propertytree as pt
-
-from ceed.propertytree.editors import PropertyEditorRegistry
-
 from ceed.editors.looknfeel.hierarchy_dock_widget import LookNFeelHierarchyDockWidget
+from ceed.editors.looknfeel.falagard_element_editor import LookNFeelFalagardElementEditorDockWidget
+from ceed.editors.looknfeel.falagard_element_inspector import FalagardElementAttributesManager
 
 
 class LookNFeelVisualEditing(QtGui.QWidget, multi.EditMode):
@@ -68,7 +61,7 @@ class LookNFeelVisualEditing(QtGui.QWidget, multi.EditMode):
 
         self.lookNFeelHierarchyDockWidget = LookNFeelHierarchyDockWidget(self, tabbedEditor)
         self.lookNFeelWidgetLookSelectorWidget = LookNFeelWidgetLookSelectorWidget(self, tabbedEditor)
-        self.lookNFeelPropertyEditorDockWidget = LookNFeelPropertyEditorDockWidget(self, tabbedEditor)
+        self.falagardElementEditorDockWidget = LookNFeelFalagardElementEditorDockWidget(self, tabbedEditor)
 
         looknfeel = QtGui.QVBoxLayout(self)
         looknfeel.setContentsMargins(0, 0, 0, 0)
@@ -82,8 +75,8 @@ class LookNFeelVisualEditing(QtGui.QWidget, multi.EditMode):
 
     def initialise(self):
         propertyMap = mainwindow.MainWindow.instance.project.propertyMap
-        widgetLookPropertyManager = CEGUIWidgetLookPropertyManager(propertyMap, self)
-        self.lookNFeelPropertyEditorDockWidget.inspector.setPropertyManager(widgetLookPropertyManager)
+        widgetLookPropertyManager = FalagardElementAttributesManager(propertyMap, self)
+        self.falagardElementEditorDockWidget.inspector.setPropertyManager(widgetLookPropertyManager)
 
         self.rootWindow = PyCEGUI.WindowManager.getSingleton().createWindow("DefaultWindow", "LookNFeelEditorRoot")
         PyCEGUI.System.getSingleton().getDefaultGUIContext().setRootWindow(self.rootWindow)
@@ -109,6 +102,10 @@ class LookNFeelVisualEditing(QtGui.QWidget, multi.EditMode):
         Destroys all child windows of the root, which means that all preview windows of the selected WidgetLookFeel should be destroyed
         :return:
         """
+
+        if self.rootWindow is None:
+            return
+
         # Remove the widget with the previous WidgetLook from the scene
         while self.rootWindow.getChildCount() != 0:
             PyCEGUI.WindowManager.getSingleton().destroyWindow(self.rootWindow.getChildAtIdx(0))
@@ -143,7 +140,7 @@ class LookNFeelVisualEditing(QtGui.QWidget, multi.EditMode):
 
         self.lookNFeelHierarchyDockWidget.setEnabled(True)
         self.lookNFeelWidgetLookSelectorWidget.setEnabled(True)
-        self.lookNFeelPropertyEditorDockWidget.setEnabled(True)
+        self.falagardElementEditorDockWidget.setEnabled(True)
 
         self.toolBar.setEnabled(True)
         if self.tabbedEditor.editorMenu() is not None:
@@ -160,7 +157,7 @@ class LookNFeelVisualEditing(QtGui.QWidget, multi.EditMode):
 
         self.lookNFeelHierarchyDockWidget.setEnabled(False)
         self.lookNFeelWidgetLookSelectorWidget.setEnabled(False)
-        self.lookNFeelPropertyEditorDockWidget.setEnabled(False)
+        self.falagardElementEditorDockWidget.setEnabled(False)
 
         self.toolBar.setEnabled(False)
         if self.tabbedEditor.editorMenu() is not None:
@@ -259,7 +256,7 @@ class LookNFeelVisualEditing(QtGui.QWidget, multi.EditMode):
         for serialisationData in topMostSerialisationData:
             serialisationData.setVisual(self)
 
-        cmd = undo.PasteCommand(self, topMostSerialisationData, targetManipulator.widget.getNamePath())
+        cmd = undoable_commands.PasteCommand(self, topMostSerialisationData, targetManipulator.widget.getNamePath())
         self.tabbedEditor.undoStack.push(cmd)
 
         # select the topmost pasted widgets for convenience
@@ -308,7 +305,7 @@ class LookNFeelWidgetLookSelectorWidget(QtGui.QDockWidget):
         selectedItemIndex = self.widgetLookNameBox.currentIndex()
         selectedWidgetLookName = self.widgetLookNameBox.itemData(selectedItemIndex)
 
-        command = undo.TargetWidgetChangeCommand(self.visual, self.tabbedEditor, selectedWidgetLookName)
+        command = undoable_commands.TargetWidgetChangeCommand(self.visual, self.tabbedEditor, selectedWidgetLookName)
         self.tabbedEditor.undoStack.push(command)
 
     def populateWidgetLookComboBox(self, widgetLookNameTuples):
@@ -331,114 +328,6 @@ class LookNFeelWidgetLookSelectorWidget(QtGui.QDockWidget):
 
         self.fileNameLabel.setToolTip(fileNameStr)
 
-
-class LookNFeelPropertyEditorDockWidget(QtGui.QDockWidget):
-    """This dock widget allows to add, remove or edit the Property, PropertyDefinition and PropertyLinkDefinition elements of a WidgetLook
-    """
-
-    def __init__(self, visual, tabbedEditor):
-        """
-        :param visual: LookNFeelVisualEditing
-        :param tabbedEditor: LookNFeelTabbedEditor
-        :return:
-        """
-        super(LookNFeelPropertyEditorDockWidget, self).__init__()
-        self.setObjectName("WidgetLookPropertiesDockWidget")
-        self.visual = visual
-
-        self.setWindowTitle("WidgetLook Properties")
-        # Make the dock take as much space as it can vertically
-        self.setSizePolicy(QtGui.QSizePolicy.Preferred, QtGui.QSizePolicy.Maximum)
-
-        self.inspector = PropertyInspectorWidget()
-        self.inspector.ptree.setupRegistry(PropertyEditorRegistry(True))
-
-        self.setWidget(self.inspector)
-
-
-class CEGUIWidgetLookPropertyManager(CEGUIWidgetLookPropertiesManager):
-    """Customises the CEGUIPropertyManager by binding to a 'visual'
-    so it can manipulate the widgets via undo commands.
-
-    It also customises the sorting of the categories.
-    """
-
-    def __init__(self, propertyMap, visual):
-        super(CEGUIWidgetLookPropertyManager, self).__init__(propertyMap)
-        self.visual = visual
-
-    def createProperty(self, ceguiProperty, ceguiSets):
-        prop = super(CEGUIWidgetLookPropertyManager, self).createProperty(ceguiProperty, ceguiSets, WidgetMultiPropertyWrapper)
-        prop.ceguiProperty = ceguiProperty
-        prop.ceguiSets = ceguiSets
-        prop.visual = self.visual
-
-        return prop
-
-    def buildCategories(self, source):
-        categories = super(CEGUIWidgetLookPropertyManager, self).buildCategories(source)
-
-        # sort categories by name but keep some special categories on top
-        def getSortKey(t):
-            name, _  = t
-
-            if name == "Element":
-                return "000Element"
-            elif name == "NamedElement":
-                return "001NamedElement"
-            elif name == "Window":
-                return "002Window"
-            elif name.startswith("CEGUI/"):
-                return name[6:]
-            elif name == "Unknown":
-                return "ZZZUnknown"
-            else:
-                return name
-
-        categories = OrderedDict(sorted(categories.items(), key=getSortKey))
-
-        return categories
-
-class WidgetMultiPropertyWrapper(pt.properties.MultiPropertyWrapper):
-    """Overrides the default MultiPropertyWrapper to update the 'inner properties'
-    and then create undo commands to update the CEGUI widgets.
-    """
-
-    def __init__(self, templateProperty, innerProperties, takeOwnership, ceguiProperty=None, ceguiSets=None, visual=None):
-        super(WidgetMultiPropertyWrapper, self).__init__(templateProperty, innerProperties, takeOwnership)
-
-        self.ceguiProperty = ceguiProperty
-        self.ceguiSets = ceguiSets
-        self.visual = visual
-
-    def tryUpdateInner(self, newValue, reason=pt.properties.Property.ChangeValueReason.Unknown):
-        if super(WidgetMultiPropertyWrapper, self).tryUpdateInner(newValue, reason):
-            ceguiValue = unicode(newValue)
-
-            # create and execute command
-            widgetPaths = []
-            undoOldValues = {}
-
-            # set the properties where applicable
-            for ceguiSet in self.ceguiSets:
-                widgetPath = ceguiSet.getNamePath()
-                widgetPaths.append(widgetPath)
-                undoOldValues[widgetPath] = self.ceguiProperty.get(ceguiSet)
-
-            # create the undoable command
-            # but tell it not to trigger the property changed callback
-            # on first run because our on value has already changed,
-            # we just want to sync the widget value now.
-            cmd = undo.PropertyEditCommand(self.visual, self.ceguiProperty.getName(), widgetPaths, undoOldValues, ceguiValue,
-                                           ignoreNextPropertyManagerCallback = True)
-            self.visual.tabbedEditor.undoStack.push(cmd)
-
-            # make sure to redraw the scene to preview the property
-            self.visual.scene.update()
-
-            return True
-
-        return False
 
 class EditingScene(cegui_widgethelpers.GraphicsScene):
     """This scene contains all the manipulators users want to interact it. You can visualise it as the
@@ -524,11 +413,11 @@ class EditingScene(cegui_widgethelpers.GraphicsScene):
                     item.preResizeSize = None
 
         if len(movedWidgetPaths) > 0:
-            cmd = undo.MoveCommand(self.visual, movedWidgetPaths, movedOldPositions, movedNewPositions)
+            cmd = undoable_commands.MoveCommand(self.visual, movedWidgetPaths, movedOldPositions, movedNewPositions)
             self.visual.tabbedEditor.undoStack.push(cmd)
 
         if len(resizedWidgetPaths) > 0:
-            cmd = undo.ResizeCommand(self.visual, resizedWidgetPaths, resizedOldPositions, resizedOldSizes, resizedNewPositions, resizedNewSizes)
+            cmd = undoable_commands.ResizeCommand(self.visual, resizedWidgetPaths, resizedOldPositions, resizedOldSizes, resizedNewPositions, resizedNewSizes)
             self.visual.tabbedEditor.undoStack.push(cmd)
 
     def keyReleaseEvent(self, event):
